feat: updating commands, queries and handlers to use my response handling instead of other responses in variants

+++ apps/merchant-api/InventoryService/src/InventoryService.Application/QueryCommands/Variants/Commands/CommandHandlers/DeleteVariantCommandHandler.cs
@@ -1,16 +1,21 @@
-using InventoryService.Application.QueryCommands.Images.Commands.Commands;
 using InventoryService.Application.QueryCommands.Variants.Commands.Commands;
+using InventoryService.Commons.ResponseHandler.Handler.Interfaces;
+using InventoryService.Commons.ResponseHandler.Responses.Bases;
 using InventoryService.Domain.Concretes;
 using InventoryService.Intraestructure.Repositories.Interfaces;
 using MediatR;
 
 namespace InventoryService.Application.QueryCommands.Variants.Commands.CommandHandlers;
 
-public class DeleteVariantCommandHandler(IRepository<Variant> variantRepository)
-    : IRequestHandler<DeleteVariantCommand, bool>
+public class DeleteVariantCommandHandler(IRepository<Variant> variantRepository, IResponseHandlingHelper responseHandlingHelper)
+    : IRequestHandler<DeleteVariantCommand, BaseResponse>
 {
-    public async Task<bool> Handle(DeleteVariantCommand request, CancellationToken cancellationToken)
+    public async Task<BaseResponse> Handle(DeleteVariantCommand request, CancellationToken cancellationToken)
     {
-        return await variantRepository.DeleteAsync(request.Id);
+        var variant = await variantRepository.GetByIdAsync(request.Id);
+        if (variant == null) return responseHandlingHelper.NotFound<Category>($"The variant with the follow id '{request.Id}' was not found.");
+
+        var response = await variantRepository.DeleteAsync(request.Id);
+        return responseHandlingHelper.Ok("The variant has been successfully deleted.", response);
     }
 }
